alzheimer_repo_runtime_download/requirements.txt
streamlit>=1.10
tensorflow>=2.13.0
numpy
Pillow
requests